requirements.txt
fastapi
uvicorn
requests
python-dotenv
sentence-transformers
qdrant-client
matplotlib